+++ app/Main.hs
@@ -2,46 +2,12 @@
 
 module Main (main) where
 
-import Prelude hiding (many, some)
-
-import Control.Monad (void, foldM)
-import Data.Maybe (listToMaybe, fromJust)
-import Data.Text (Text)
-import qualified Data.Text as T
 import qualified Data.Text.IO as TIO
 import System.Environment (getArgs)
-import Text.Pandoc.Class (PandocIO, PandocMonad, runIOorExplode)
-import Text.Pandoc.Definition (Pandoc)
-import Text.Pandoc.Options (ReaderOptions, def)
-import Text.Pandoc.Readers (Reader (TextReader), readers)
-import Text.Megaparsec (ParsecT, runParserT, (<?>), many, some, errorBundlePretty)
-import qualified Text.Megaparsec.Char as MPC
-import Control.Monad.Trans.State (State, evalState)
-import Data.Void (Void)
+import Text.Megaparsec (runParserT, errorBundlePretty)
+import Control.Monad.Trans.State (evalState)
 
-type TxtReader = ReaderOptions -> Text -> PandocIO Pandoc
-
--- Gets the reader associated with a name (Ex: "markdown")
-getPandocTxtReader :: Text -> Maybe TxtReader
-getPandocTxtReader k =
-  case listToMaybe $ fmap snd $ filter ((== k) . fst) readers of
-    Just (TextReader r) -> Just r
-    -- Ignores bytestring readers
-    _ -> Nothing
-
-data BoarConfig = BoarConfig { markup :: Maybe (Text, TxtReader)
-                             }
-defaultBoarConfig :: BoarConfig
-defaultBoarConfig = BoarConfig { markup = Nothing
-                               }
-
-
-
-setMarkupPrefix :: Text
-setMarkupPrefix = "boar set markup "
-
-
-
+import FParser (defaultFState, parseFile)
 
 singleArgument :: [String] -> String
 singleArgument [arg] = arg
@@ -56,19 +22,5 @@ main = do
     Right x -> print x
     Left e -> putStrLn $ errorBundlePretty e
 
-data FState = FState { bCfg :: BoarConfig
-                     }
-defaultFState :: FState
-defaultFState = FState { bCfg = defaultBoarConfig
-                       }
-
-type FParser = ParsecT Void Text (State FState)
-
-parseFile :: FParser ()
-parseFile =
-  do
-    _ <- MPC.char 'x'
-    return ()
-  <?> "x"
